+++ src/main/java/com/paradigma0621/report/dto/ClockingRegisterPublisherProtocolDto.java
@@ -1,0 +1,3 @@
+package com.paradigma0621.report.dto;
+
+public record ClockingRegisterPublisherProtocolDto(String protocolo){}