+++ src/main/java/com/paradigma0621/report/exception/ClockingSenderRequestException.java
@@ -1,0 +1,13 @@
+package com.paradigma0621.report.exception;
+
+public class ClockingSenderRequestException extends RuntimeException {
+
+    public ClockingSenderRequestException() {
+        super();
+    }
+
+    public ClockingSenderRequestException(String message) {
+        super(message);
+    }
+
+}